Add specific database schema blueprint classes to rdbms
